mongo_to_snowflake/config.py
import os
from dotenv import load_dotenv

# Load .env file
load_dotenv()

# --- MongoDB Configuration ---
MONGO_URI = os.getenv("MONGO_URI")
MONGO_DB_NAME = os.getenv("MONGO_DB_NAME", "sample_training")
MONGO_COLLECTION_NAME = os.getenv("MONGO_COLLECTION_NAME", "grades")

# --- Snowflake Configuration ---
SNOWFLAKE_USER = os.getenv("SNOWFLAKE_USER")
SNOWFLAKE_PASSWORD = os.getenv("SNOWFLAKE_PASSWORD")
SNOWFLAKE_ACCOUNT = os.getenv("SNOWFLAKE_ACCOUNT")
SNOWFLAKE_WAREHOUSE = os.getenv("SNOWFLAKE_WAREHOUSE")
SNOWFLAKE_DATABASE = os.getenv("SNOWFLAKE_DATABASE")
SNOWFLAKE_SCHEMA = os.getenv("SNOWFLAKE_SCHEMA")
SNOWFLAKE_TABLE = os.getenv("SNOWFLAKE_TABLE")
SNOWFLAKE_PSA_SCHEMA = os.getenv("SNOWFLAKE_PSA_SCHEMA")
SNOWFLAKE_PSA_TABLE= os.getenv("SNOWFLAKE_PSA_TABLE")