requirements.txt
pymongo[srv]==3.11
snowflake-connector-python==3.0.3
python-dotenv==1.1.1
PyYAML==6.0.2